requirements.txt
diffdist==0.1
numpy==1.21.6
Pillow==10.2.0
scikit_learn==1.0.2
torch==1.12.1
torchvision==0.13.1
pandas==1.3.5
tqdm==4.64.1
timm==0.6.7
PyYAML==6.0.2
regex==2024.11.6
ftfy==6.3.1
continuum==1.2.7